requirements.txt
flask
gunicorn
numpy
pandas
base64
matplotlib.pyplot
pandas_datareader
yfinance
keras.models
sklearn.preprocessing